typecast x and y
the last commit only cast the result of dividing two longs to be a
float, this commit exhibits c typecasting syntax for x and y as well

+++ cs50/calculator.c
@@ -6,6 +6,6 @@ int main(void)
     long x = get_long("x: ");
     long y = get_long("y: ");
 
-    float z = x / y;
+    float z = (float) x / (float) y;
     printf("%f\n", z);
 }
